+++ src/main/java/com/example/bank/BankApplication.java
@@ -2,10 +2,12 @@ package com.example.bank;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 import java.util.TimeZone;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class BankApplication {
 
     public static void main(String[] args) {